Fix edge case with threads getting scheduled with old state.
See comment in schedule.c for details. This could happen during
a thread exiting while another thread unblocks it. The slab
allocator could give us back a new thread that is still on a
runqueue, letting a thread run with old state. This is bad.

Fix this by having the scheduler not run a thread that is being
initialized if it happens to find one. While unlikely, this does
happen in testing.