Implement directory cache, and fix resource leak.
The VFS system does lazy clean up, with one important point. When
a call to unlink is made, the dirent gets a flag that says that it
was unlinked. Then, when dirent_put is called, the function checks
the flag and removes it from the LRU cache if it is set (because
there is no point keeping it around if its been unlinked, and
because we want to actually decrement the link count for the inode
as soon as we can. Once the dirent put function is called, if we've
unlinked it, actually do the unlink and decrement the link count.
This makes checking if a directory is empty a bit harder, since we
have to check the dirents in the cache to make sure that nothing
funny is going on.

The same logic applies to inodes: if their link count is zero when
inode_put is called, remove it from the cache and such. Actually
do the inode release when inode's put function is called.