Fix a corrupted data bug in kobj_lru.
When an object is reclaimed, there was a window where a thread
could pull in a new copy of the object as it was being released.
This could result in two different copies (one with potentially
old data) in the system at once, which is Bad (tm).

This fixes it by adding a new flag. Though the kobj_lru code really
needs some cleanup now...