Set PCI interrupts to level-triggered.
PCI devices that share an interrupt line need level-triggered interrupts
in case one device causes an interrupt, and the other then causes it
before the ACK. This would be a race condition in the case of edge
triggered.

This commit solves the lost AHCI interrupt problem after enabling
the e1000 device.